requirements.txt
fastapi==0.103.2
keras==2.14.0
numpy==1.23.5
Pillow==9.4.0
Pillow==10.0.1
pydantic==1.10.8
Requests==2.31.0
